OurArticle/Article/requirement.txt
pillow
django-ckeditor